Backend/requirements.txt
flask == 1.1.2
redis
pymysql
pymongo
flask-mysql
jsonify
requests

gunicorn
bcrypt
PyJWT
flask_cors
flask_restx
werkzeug

tensorflow==2.2
keras==2.4.3 
numpy==1.19.3 
pillow==7.0.0 
scipy==1.4.1 
h5py==2.10.0 
matplotlib==3.3.2 
opencv-python 
keras-resnet==0.2.0
imageai 
awscli
tesseract
tesseract-ocr
pytesseract
image


boto3

